--- src/extension/src/utils/dependencyChecker/IDependencyChecker.ts
@@ -1,3 +1,0 @@
-interface IDependencyChecker {
-  hasDependency: () => Promise<boolean>;
-}
