--- src/extension/src/utils/dependencyChecker/checkers/netCoreChecker.ts
@@ -1,19 +1,0 @@
-import util = require("util");
-import semver = require("semver");
-
-export default class NetCoreChecker implements IDependencyChecker {
-    private exec = util.promisify(require("child_process").exec);
-    private NETCORE_REGEX = RegExp("([0-9.]+)");
-    private NETCORE_REQUIREMENT = ">=3.1.x";
-
-    public async hasDependency() {
-        try {
-            const { stdout } = await this.exec("dotnet --version");
-            const version = stdout.match(this.NETCORE_REGEX)[1];
-            const result = semver.satisfies(version, this.NETCORE_REQUIREMENT);
-            return result;
-          } catch (err) {
-            return false;
-          }
-    }
-  }
