+++ src/extension/src/utils/requirements/IRequirementValidator.ts
@@ -1,0 +1,3 @@
+interface IRequirementValidator {
+  isInstalled: () => Promise<boolean>;
+}
