+++ src/extension/src/utils/requirements/validators/nodeValidator.ts
@@ -2,13 +2,12 @@ import util = require("util");
 import semver = require("semver");
 import { CONSTANTS } from "../../../constants";
 
-export default class NodeChecker implements IDependencyChecker {  
+export default class NodeValidator implements IRequirementValidator {
+  private exec = util.promisify(require("child_process").exec);
+  private NODE_REGEX = RegExp("v(.+)");
+  private NODE_REQUIREMENT = ">=12.0.x";
 
-private exec = util.promisify(require("child_process").exec);
-private NODE_REGEX = RegExp("v(.+)");
-private NODE_REQUIREMENT = ">=12.0.x";
-
-  public async hasDependency() {
+  public async isInstalled() {
     try {
       const { stdout } = await this.exec(CONSTANTS.DEPENDENCY_CHECKER.NODE + " --version");
       const version = stdout.match(this.NODE_REGEX)[1];
